Fix signed / unsigne comparison.

+++ src/isxMosaicMovieFile.cpp
@@ -109,7 +109,7 @@ MosaicMovieFile::writeFrame(const SpVideoFrame_t & inVideoFrame)
 
     std::ofstream file(m_fileName, std::ios::binary | std::ios::app);
 
-    if (currentFileSize != int64_t(inVideoFrame->getFrameIndex() * getFrameSizeInBytes() + m_headerOffset))
+    if (currentFileSize != inVideoFrame->getFrameIndex() * getFrameSizeInBytes() + m_headerOffset)
     {
         ISX_LOG_ERROR("MosaicMovieFile::writeFrame: Attempt to write frames out of order.");
         ISX_ASSERT(false);
